+++ src/main/java/fr/eni/buymystuff/jdbc/IDAOCategories.java
@@ -7,11 +7,10 @@ import java.util.List;
 public interface IDAOCategories {
 
     public List<Categories> selectAllCategories();
-    //public Categories selectCategoryById(int id);
     public Categories save(Categories categories);
     public Categories selectByLibelle(String libelle);
 
-    Categories selectById(int id);
+    Categories selectById(Long id);
     //public Categories update(Categories categories);
     //public Categories delete(Categories categories);
 }
